old-tests/webapps/XMLHttpRequest/tests/submissions/Microsoft/send-network-error-async-events.htm

<!DOCTYPE html>
<html>
<head>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-send()-method" />
    <script src="/resources/testharness.js"></script>
    <title>XmlHttpRequest L2: The send() method: Fire a progress event named error when Network error happens (synchronous flag is unset)</title>
</head>

<body>
    <div id="log"></div>

    <script type="text/javascript">
        var test = async_test();

        test.step(function(){
            var xhr = new XMLHttpRequest();
            var expect =  [4, "upload.error", "upload.loadend", "error", "loadend"];
            var actual = [];

            xhr.onreadystatechange = function()
            {
                test.step(function()
                {
                    if (xhr.readyState == 4) 
                    {
                        actual.push(xhr.readyState);
                    }
                });
            };

            xhr.onloadend          = function(e){ actual.push(e.type); VerifyResult()};
            xhr.onerror            = function(e){ actual.push(e.type); };
            
            xhr.upload.onloadend   = function(e){ actual.push("upload." + e.type);};
            xhr.upload.onerror     = function(e){ actual.push("upload." + e.type); };

            function VerifyResult()
            {
                test.step(function()
                {
                    assert_array_equals(actual, expect);
                    test.done();
                });
            };

            xhr.open("POST", "http://nonexistent-origin.w3c-test.org", true);
            xhr.send("Test Message");
        });
    </script>
</body>
</html>
